Customer_commit_3.

+++ property-management-backend/src/main/java/com/waa/PropertyManagment/service/SavedListService.java
@@ -8,6 +8,7 @@ import com.waa.PropertyManagment.repository.SavedListRepository;
 import com.waa.PropertyManagment.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
@@ -32,15 +33,46 @@ public class SavedListService {
         if(savedListOptional.isEmpty()){
             SavedList savedList=new SavedList();
             savedList.setName(name);
-            savedList.setProperties(property);
+            savedList.setProperties((Arrays.asList(property)));
             savedList.setCustomer(user);
             savedListRepository.save(savedList);
         }
     }
+
+
+/*
+
+    public SavedList getSavedList(Long id) {
+        return savedListRepository.findById(id).orElse(null);
+    }
+    public SavedList addPropertyToSavedList(Long savedListId, Long propertyId) {
+        SavedList savedList = getSavedList(savedListId);
+        if (savedList == null) {
+            return null;
+        }
+
+        Property property = propertyRepository.findById(propertyId).orElse(null);
+        if (property == null) {
+            return null;
+        }
+
+        property.setSavedList(savedList);
+        propertyRepository.save(property);
+        return savedList;
+    }
+
+*/
+
     public List<SavedList> getSavedProperties(Long customerId){
         return savedListRepository.findByCustomerId(customerId);
     }
     public void removeSavedProperty(Long savedPropertyId){
-        savedListRepository.deleteById(savedPropertyId);
+        if (savedListRepository.findById(savedPropertyId).isPresent()) {
+            savedListRepository.deleteById(savedPropertyId);
+        }
+        else {
+            throw  new RuntimeException("Property Id not found");
+        }
     }
+
 }
